+++ web_app/app.py
@@ -1,7 +1,8 @@
+import os
 from flask import Flask, render_template, request
-import loading_dataset as ld
-from CNN_model import MLP_model
-import torch
+import numpy as np
+from datetime import datetime
+
 
 app = Flask(__name__)
 
@@ -11,11 +12,22 @@ def index():
 
 @app.route("/receive",methods=["POST"])
 def receive_data():
-    
     data = request.get_json()
-    pythonConvert = data.get('pythonConvert', [])
-    print(pythonConvert)
-    return 'Data received successfully'
+    image_data = data.get('imageData', [])
+    print(len(image_data))
+    image_data = np.asarray(image_data)
+    
+    # For accessing the color array outside, you could save it to a file or pass it to another function
+    image_data = image_data.reshape((512, 512, 4))[:, :, :3]
+    timestamp = datetime.now().strftime("%Y%m%d-%H%M%S")
+    directory_path = 'images/saved_arrays'
+    os.makedirs(directory_path, exist_ok=True)
+    filename = f"{directory_path}/image_data_{timestamp}.npy"
+    
+    # Save the array to the file
+    np.save(filename, image_data)
+    letter = "A"
+    return {'letter': letter}
 
 if __name__ == "__main__":
-    app.run(port=8000,debug=True)
+    app.run(port=5000,debug=True)
