+++ web_app/templates/index.html
@@ -5,8 +5,8 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>MAIS-202-F2023-ASL-PROCESSING</title>
     <link rel="icon" href="../images/logo-icon.png">
-    <link rel="stylesheet" href="../style.css">
-    <script src="../main.js" defer></script>
+    <link rel="stylesheet" href="{{ url_for('static', filename='style.css') }}">
+    <script src="{{ url_for('static', filename='main.js') }}" defer></script>   
 </head>
 <body>
     <header>
